combo box categoria
tmb
-boton editar en tecnicos
-lista de categorias

+++ src/main/java/com/sebaandino/tpi/Persistencia/TecnicoJpaController.java
@@ -60,7 +60,7 @@ public class TecnicoJpaController implements Serializable {
         } catch (Exception ex) {
             String msg = ex.getLocalizedMessage();
             if (msg == null || msg.length() == 0) {
-                Long id = tecnico.getIdTecnico();
+                int id = tecnico.getIdTecnico();
                 if (findTecnico(id) == null) {
                     throw new NonexistentEntityException("The tecnico with id " + id + " no longer exists.");
                 }
@@ -73,7 +73,7 @@ public class TecnicoJpaController implements Serializable {
         }
     }
 
-    public void destroy(Long id) throws NonexistentEntityException {
+    public void destroy(int id) throws NonexistentEntityException {
         EntityManager em = null;
         try {
             em = getEntityManager();
@@ -118,7 +118,7 @@ public class TecnicoJpaController implements Serializable {
         }
     }
 
-    public Tecnico findTecnico(Long id) {
+    public Tecnico findTecnico(int id) {
         EntityManager em = getEntityManager();
         try {
             return em.find(Tecnico.class, id);
